Fix bug with missing URL search params
