+++ src/helpers/hooks.js
@@ -1,5 +1,4 @@
 import { useState, useEffect } from 'react';
-import { useHistory, useLocation } from 'react-router-dom';
 import { getPeople } from '../api';
 
 export function usePeople() {
@@ -21,22 +20,4 @@ export function usePeople() {
   }, []);
 
   return [people, setPeople];
-}
-
-export function useSearchParams() {
-  const history = useHistory();
-  const location = useLocation();
-  const searchParams = new URLSearchParams(location.search);
-
-  const updateSearchParams = (key, value) => {
-    if (value) {
-      searchParams.set(key, value);
-    } else {
-      searchParams.delete(key);
-    }
-
-    history.push({ search: searchParams.toString() });
-  };
-
-  return [searchParams, updateSearchParams];
 }
